+++ src/state-priv.h
@@ -152,37 +152,70 @@ assert_same_field(struct xkb_state_update, _v1, layout_policy);
 static_assert(sizeof(struct xkb_state_update) * 30 <=
               (size_t)XKB_ABI_MAX_SIZE, "");
 
+/**
+ * Version 1 of `xkb_machine_builder_a11y_update`, used for ABI check only
+ *
+ * @since 1.14.0
+ */
+struct xkb_machine_builder_a11y_update_v1 {
+    uint32_t size;
+    uint32_t affect;
+    uint32_t flags;
+};
+
+/* Ensure there is no implicit padding */
+assert_no_padding(struct xkb_machine_builder_a11y_update, size, affect);
+assert_no_padding(struct xkb_machine_builder_a11y_update, affect, flags);
+assert_no_padding(struct xkb_machine_builder_a11y_update, flags);
+
+/* Current version is 1 */
+static_assert(sizeof(struct xkb_machine_builder_a11y_update) ==
+              sizeof(struct xkb_machine_builder_a11y_update_v1), "");
+assert_same_field(struct xkb_machine_builder_a11y_update, _v1, size);
+assert_same_field(struct xkb_machine_builder_a11y_update, _v1, affect);
+assert_same_field(struct xkb_machine_builder_a11y_update, _v1, flags);
+
+/* Ensure reasonable margin to the upper size limit */
+static_assert(sizeof(struct xkb_machine_builder_a11y_update) * 30 <=
+              (size_t)XKB_ABI_MAX_SIZE, "");
+
 /** Size of the *first version* of the struct */
-#define xkb_versioned_struct_size_v1(x) _Generic(      \
-    (x),                                               \
-    const struct xkb_state_update *:                   \
-        sizeof(struct xkb_state_update_v1),            \
-    const struct xkb_state_components_update *:        \
-        sizeof(struct xkb_state_components_update_v1), \
-    const struct xkb_layout_policy_update *:           \
-        sizeof(struct xkb_layout_policy_update_v1)     \
+#define xkb_versioned_struct_size_v1(x) _Generic(                 \
+    (x),                                                          \
+    const struct xkb_state_update *:                              \
+        sizeof(struct xkb_state_update_v1),                       \
+    const struct xkb_state_components_update *:                   \
+        sizeof(struct xkb_state_components_update_v1),            \
+    const struct xkb_layout_policy_update *:                      \
+        sizeof(struct xkb_layout_policy_update_v1),               \
+    const struct xkb_machine_builder_a11y_update *:               \
+        sizeof(struct xkb_machine_builder_a11y_update_v1)         \
 )
 
 /** Minimal *current* valid size of the struct */
-#define xkb_versioned_struct_size_min(x) _Generic(     \
-    (x),                                               \
-    const struct xkb_state_update *:                   \
-        sizeof(struct xkb_state_update_v1),            \
-    const struct xkb_state_components_update *:        \
-        sizeof(struct xkb_state_components_update_v1), \
-    const struct xkb_layout_policy_update *:           \
-        sizeof(struct xkb_layout_policy_update_v1)     \
+#define xkb_versioned_struct_size_min(x) _Generic(                \
+    (x),                                                          \
+    const struct xkb_state_update *:                              \
+        sizeof(struct xkb_state_update_v1),                       \
+    const struct xkb_state_components_update *:                   \
+        sizeof(struct xkb_state_components_update_v1),            \
+    const struct xkb_layout_policy_update *:                      \
+        sizeof(struct xkb_layout_policy_update_v1),               \
+    const struct xkb_machine_builder_a11y_update *:               \
+        sizeof(struct xkb_machine_builder_a11y_update_v1)         \
 )
 
 /** Offset of the last meaningful field of the current struct */
-#define xkb_versioned_struct_reserved_offset(x) _Generic( \
-    (x),                                                  \
-    const struct xkb_state_update *:                      \
-        sizeof(struct xkb_state_update),                  \
-    const struct xkb_state_components_update *:           \
-        sizeof(struct xkb_state_components_update),       \
-    const struct xkb_layout_policy_update *:              \
-        sizeof(struct xkb_layout_policy_update)           \
+#define xkb_versioned_struct_reserved_offset(x) _Generic(        \
+    (x),                                                         \
+    const struct xkb_state_update *:                             \
+        sizeof(struct xkb_state_update),                         \
+    const struct xkb_state_components_update *:                  \
+        sizeof(struct xkb_state_components_update),              \
+    const struct xkb_layout_policy_update *:                     \
+        sizeof(struct xkb_layout_policy_update),                 \
+    const struct xkb_machine_builder_a11y_update *:              \
+        sizeof(struct xkb_machine_builder_a11y_update)           \
 )
 
 #define xkb_check_update_size(x) xkb_check_versioned_struct_size( \
@@ -208,6 +241,11 @@ static_assert(
     xkb_versioned_struct_size_min(((const struct xkb_layout_policy_update *)NULL)),
     ""
 );
+static_assert(
+    xkb_versioned_struct_size_v1(((const struct xkb_machine_builder_a11y_update *)NULL)) <=
+    xkb_versioned_struct_size_min(((const struct xkb_machine_builder_a11y_update *)NULL)),
+    ""
+);
 
 /* Minimal size is lower or equal to the current size */
 static_assert(
@@ -223,6 +261,11 @@ static_assert(
 static_assert(
     xkb_versioned_struct_size_min(((const struct xkb_layout_policy_update *)NULL)) <=
     sizeof(const struct xkb_layout_policy_update),
+    ""
+);
+static_assert(
+    xkb_versioned_struct_size_min(((const struct xkb_machine_builder_a11y_update *)NULL)) <=
+    sizeof(const struct xkb_machine_builder_a11y_update),
     ""
 );